# Conflicts:
#	.classpath
#	src/org/sosy_lab/cpachecker/cfa/ast/AFloatLiteralExpression.java
#	src/org/sosy_lab/cpachecker/cfa/ast/c/CFloatLiteralExpression.java
#	src/org/sosy_lab/cpachecker/cfa/ast/java/JFloatLiteralExpression.java
#	src/org/sosy_lab/cpachecker/cpa/smg2/SMGCPA.java
#	src/org/sosy_lab/cpachecker/cpa/smg2/SMGCPAValueVisitor.java
#	src/org/sosy_lab/cpachecker/cpa/value/AbstractExpressionValueVisitor.java
#	src/org/sosy_lab/cpachecker/cpa/value/ValueAnalysisState.java
#	src/org/sosy_lab/cpachecker/util/floatingpoint/CFloatImpl.java
#	src/org/sosy_lab/cpachecker/util/floatingpoint/NativeComputationException.java

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/fix_floatingpoint_calculator@47416 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cfa/simplification/ExpressionSimplificationVisitor.java
@@ -286,7 +286,7 @@ public class ExpressionSimplificationVisitor
             return new CFloatLiteralExpression(
                 loc, machineModel, exprType, negatedValue.floatingPointValue());
           default:
-            // fall-through and return the original expression
+            // return the original expression below
         }
 
       } else if (unaryOperator == UnaryOperator.TILDE
@@ -332,8 +332,8 @@ public class ExpressionSimplificationVisitor
 
     // enum constant
     if (decl instanceof CEnumerator) {
-      final long v = ((CEnumerator) decl).getValue();
-      return new CIntegerLiteralExpression(expr.getFileLocation(), type, BigInteger.valueOf(v));
+      final BigInteger v = ((CEnumerator) decl).getValue();
+      return new CIntegerLiteralExpression(expr.getFileLocation(), type, v);
     }
 
     // const variable, inline initializer
@@ -361,7 +361,7 @@ public class ExpressionSimplificationVisitor
               return new CFloatLiteralExpression(
                   expr.getFileLocation(), machineModel, type, castValue);
             default:
-              // fall-through and return the original expression
+              // return the original expression below
           }
         }
       }
